Небольшие исправления
